# contracts/api/openapi.json
{
  "components": {
    "schemas": {
...
      "Body_upload_dataset_api_v1_datasets_upload_post": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          },
          "submitted_by": {
            "title": "Submitted By",
            "type": "string"
          },
          "topic_hint": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Topic Hint"
          }
        },
        "required": [
          "file",
          "submitted_by"
        ],
        "title": "Body_upload_dataset_api_v1_datasets_upload_post",
        "type": "object"
      },
...
      "CompleteUploadRequest": {
        "description": "The browser callback after the multipart form succeeds.",
        "properties": {
          "objectKey": {
            "minLength": 1,
            "title": "Objectkey",
            "type": "string"
          },
          "topicHint": {
            "anyOf": [
              {
                "maxLength": 100,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Topichint"
          }
        },
        "required": [
          "objectKey"
        ],
        "title": "CompleteUploadRequest",
        "type": "object"
      },
...
      "FileFormat": {
        "enum": [
          "csv",
          "excel",
          "json",
          "pdf"
        ],
        "title": "FileFormat",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "JobStatus": {
        "description": "Lifecycle of an ingestion job.",
        "enum": [
          "pending",
          "running",
          "awaiting_approval",
          "published",
          "quarantined",
          "rejected",
          "failed"
        ],
        "title": "JobStatus",
        "type": "string"
      },
      "JobStatusResponse": {
        "properties": {
          "createdAt": {
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "currentStep": {
            "title": "Currentstep",
            "type": "string"
          },
          "datasetId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Datasetid"
          },
          "jobId": {
            "title": "Jobid",
            "type": "string"
          },
          "links": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Links",
            "type": "object"
          },
          "qualityScore": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Qualityscore"
          },
          "sourceFormat": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FileFormat"
              },
              {
                "type": "null"
              }
            ]
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "jobId",
          "status",
          "currentStep",
          "createdAt"
        ],
        "title": "JobStatusResponse",
        "type": "object"
      },
...
      "StartedIngestionResponse": {
        "description": "Portable reference to the workflow started for an uploaded object.",
        "properties": {
          "createdAt": {
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "jobId": {
            "title": "Jobid",
            "type": "string"
          },
          "links": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Links",
            "type": "object"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus"
          }
        },
        "required": [
          "jobId",
          "status",
          "createdAt",
          "links"
        ],
        "title": "StartedIngestionResponse",
        "type": "object"
      },
...
      "UploadRequest": {
        "description": "A request for a presigned upload URL.",
        "properties": {
          "contentType": {
            "examples": [
              "text/csv"
            ],
            "title": "Contenttype",
            "type": "string"
          },
          "originalFilename": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "examples": [
              "employment_2024.csv"
            ],
            "title": "Originalfilename"
          },
          "submittedBy": {
            "examples": [
              "steward@ntpc.gov.tw"
            ],
            "maxLength": 256,
            "minLength": 1,
            "title": "Submittedby",
            "type": "string"
          }
        },
        "required": [
          "contentType",
          "submittedBy"
        ],
        "title": "UploadRequest",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "apps__api__schemas__UploadResponse": {
        "properties": {
          "jobId": {
            "title": "Jobid",
            "type": "string"
          },
          "links": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Links",
            "type": "object"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus"
          }
        },
        "required": [
          "jobId",
          "status",
          "links"
        ],
        "title": "UploadResponse",
        "type": "object"
      },
      "apps__api__uploads__UploadResponse": {
        "description": "A presigned POST the browser submits the file to.",
        "properties": {
          "expiresAt": {
            "format": "date-time",
            "title": "Expiresat",
            "type": "string"
          },
          "fields": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Fields",
            "type": "object"
          },
          "jobId": {
            "title": "Jobid",
            "type": "string"
          },
          "links": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Links",
            "type": "object"
          },
          "maxBytes": {
            "title": "Maxbytes",
            "type": "integer"
          },
          "objectKey": {
            "title": "Objectkey",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus"
          },
          "url": {
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "url",
          "fields",
          "jobId",
          "objectKey",
          "status",
          "expiresAt",
          "maxBytes",
          "links"
        ],
        "title": "UploadResponse",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "New Taipei Youth Policy API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/datasets/upload": {
      "post": {
        "operationId": "upload_dataset_api_v1_datasets_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_dataset_api_v1_datasets_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/apps__api__schemas__UploadResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload Dataset",
        "tags": [
          "reviewer"
        ]
      }
    },
...
    "/api/v1/ingestion-jobs/{job_id}": {
      "get": {
        "operationId": "get_ingestion_job_api_v1_ingestion_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Ingestion Job",
        "tags": [
          "reviewer"
        ]
      }
    },
...
    "/api/v1/uploads": {
      "post": {
        "description": "Issue a presigned POST for one file upload into the incoming zone.",
        "operationId": "create_upload_api_v1_uploads_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UploadRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/apps__api__uploads__UploadResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Upload",
        "tags": [
          "uploads"
        ]
      }
    },
    "/api/v1/uploads/{job_id}/complete": {
      "post": {
        "description": "Verify the S3 object and start one idempotently named workflow.",
        "operationId": "complete_upload_api_v1_uploads__job_id__complete_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompleteUploadRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StartedIngestionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Complete Upload",
        "tags": [
          "uploads"
        ]
      }
    },
...
  }
}